% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc.data2sf.R
\name{arc.data2sf}
\alias{arc.data2sf}
\title{Convert an arc.dataframe object to an sf Simple Feature object}
\usage{
arc.data2sf(x)
}
\arguments{
\item{x}{\code{data.frame} result of \code{\link{arc.select}}}
}
\description{
Convert an ArcGIS \code{data.frame} to the equivalent \code{sf} object
type. The output types that can be generated:
\code{POINT}, \code{MULTIPOINT}, \code{POLYGON}, \code{MULTIPOLYGON}, \code{LINESTRING}, \code{MULTILINESTRING}.
}
\examples{

require(sf)
d <- arc.select(arc.open(system.file("extdata", "ca_ozone_pts.shp", package="arcgisbinding")), 'ozone')
x <- arc.data2sf(d)
\dontrun{plot(x)}
}

