% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA.classes.R
\docType{class}
\name{arc.datasetraster-class}
\alias{arc.datasetraster-class}
\alias{dim,arc.datasetraster-method}
\alias{names,arc.datasetraster-method}
\title{Class "arc.datasetraster"}
\description{
\code{arc.datasetraster} S4 class
}
\details{
TODO
}
\section{Slots}{

\describe{
\item{\code{.info}}{internal}

\item{\code{sr}}{spatial reference}

\item{\code{extent}}{spatial extent of the dataset}

\item{\code{pixel_type}}{The pixel type of the referenced raster dataset. (\link[=pixeltypes]{pixel_type})}

\item{\code{compression_type}}{The compression type}

\item{\code{nrow}}{return the number of rows}

\item{\code{ncol}}{return the number of rows}

\item{\code{bands}}{return rasterdataset bands information}
}}
\section{Methods}{

\describe{
  \item{\link{arc.raster}}{}
  \item{\link{dim}}{}
  \item{\link{names}}{}
}
}
\seealso{
\code{\link{arc.open}}
}
\keyword{dataset}
\keyword{open}
\keyword{raster}
\keyword{rasterdataset}

