\name{arc.metadata}
\docType{methods}
\alias{arc.metadata}
\alias{arc.metadata,arc.container-method}
\alias{arc.metadata,arc.dataset-method}
\title{Get metadata}
\usage{
  \S4method{arc.metadata}{arc.dataset}(object)
}
\arguments{
\item{object}{\link{arc.dataset-class} object}
}
\description{
Get metadata
}

\keyword{internal}

