% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/A.arc.classes.R
\docType{class}
\name{arc.shape-class}
\alias{arc.shape-class}
\alias{arc.shapeinfo,arc.shape-method}
\alias{show,arc.shape-method}
\title{Class "arc.shape"}
\description{
\code{arc.shape} is geometry collection
}
\note{
\code{arc.shape} is attached to an ArcGIS \code{data.frame} as the
attribute "shape". Each element corresponds to one record in
the input data frame. Points are presented as an array of lists, with
each list containing (\code{x}, \code{y}, \code{Z}, \code{M}), where
}
\examples{
d <- arc.open(system.file("extdata", "ca_ozone_pts.shp",
                          package="arcgisbinding"))
arc.df <- arc.select(d, "FID")

shape <- arc.shape(arc.df)
# create data.frame with X and Y columns
df <- data.frame(arc.df, X=shape$x, Y=shape$y)
# print out row #42
df[42,]
}
\keyword{classes}
\keyword{geometry}
\keyword{shape}

