% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc.data2sp.R
\name{arc.shape2sp}
\alias{arc.shape2sp}
\title{Convert Esri shape to sp spatial geometry}
\usage{
arc.shape2sp(shape, wkt)
}
\arguments{
\item{shape}{\code{\link{arc.shape-class}}}

\item{wkt}{WKT spatial reference}
}
\description{
Convert \code{\link{arc.shape-class}} to \code{sp} spatial geometry:
\code{SpatialPoints}, \code{SpatialLines}, or \code{SpatialPolygons}.
}
\examples{

require(sp)
d <- arc.select(arc.open(system.file("extdata", "ca_ozone_pts.shp", package="arcgisbinding")), 'ozone')
x <- arc.shape2sp(arc.shape(d))
\dontrun{plot(x)}
}
\seealso{
\code{\link{arc.shape}}
}

