\name{arc.sp2data}
\alias{arc.sp2data}
\title{deprecated Convert a sp SpatialDataFrame object to an arc.dataframe object}
\arguments{
\item{sp.df}{\code{SpatialPointsDataFrame}, \code{SpatialPolygonsDataFrame},
or \code{SpatialLinesDataFrame}}
}
\description{
deprecated Convert \code{sp} \code{SpatialPointsDataFrame}, \code{SpatialPolygonsDataFrame},
and \code{SpatialLinesDataFrame} objects to an ArcGIS-compatible \code{data.frame}.
}
\seealso{
\code{\link{arc.data2sp}}
}

\keyword{internal}
