% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.generics.R
\name{arc.write}
\alias{arc.write}
\title{Write dataset, raster, feature, table or layer}
\arguments{
\item{path}{full output path}

\item{data}{input source. Accepts \code{data.frame}, spatial \code{data.frame},
\code{SpatialPointsDataFrame}, \code{SpatialLinesDataFrame}, and
\code{SpatialPolygonsDataFrame}, \code{arc.raster}, \code{raster::RasterLayer}, \code{raster::RasterBrick} objects.}

\item{\dots}{Optional parameters
 \itemize{
   \item \code{coords} list containing geometry. Accepts \code{Spatial} objects. Put field names if \code{data} is data.frame and consists coordinates.
   \item \code{shape_info} required argument if \code{data} has no spatial attribute
   \item \code{overwrite} overwrite existing dataset. default = FALSE.
}}
}
\description{
Export a \code{data.frame} object to an ArcGIS dataset. If the data frame
includes a spatial attribute, this function writes a feature dataset. If no
spatial attribute is found, a table is instead written.\cr
Export a \code{arc.raster}, \code{raster::RasterLayer} or \code{raster::RasterBrick} object to an ArcGIS raster dataset.
}
\details{
Supports a variety of output formats. Below are pairs of example paths and the resulting data types:
 \itemize{
  \item \code{C:/place.gdb/fc}: File Geodatabase Feature Class
  \item \code{C:/place.gdb/fdataset/fc}: File Geodatabase Feature Dataset
  \item \code{in_memory\\logreg}: In-memory workspace (must be run in ArcGIS Session)
  \item \code{C:/place.shp}: Esri Shapefile
  \item \code{C:/place.dbf}: Table
  \item \code{C:/place.gdb/raster}: File Geodatabase Raster when \code{data} parameter is \code{arc.raster} or \code{Raster*} object
  \item \code{C:/image.img}: ERDAS Imaging
  \item \code{C:/image.tif}: Geo TIFF
}
}
\section{References}{

\itemize{
\item \href{http://support.esri.com/es/knowledgebase/techarticles/detail/40057}{What is the difference between a shapefile and a layer file?}
\item \href{https://desktop.arcgis.com/en/desktop/latest/map/working-with-layers/what-is-a-layer-.htm}{ArcGIS Help: What is a layer?}
}
}
\examples{

## write as a shapefile
fc <- arc.open(system.file("extdata", "ca_ozone_pts.shp", package="arcgisbinding"))
d <- arc.select(fc, 'ozone')
d[1,] <- 0.6
arc.write(tempfile("ca_new", fileext=".shp"), d)

## create and write to a new file geodatabase
fgdb_path <- file.path(tempdir(), "data.gdb")

data(meuse, package="sp")
## create feature dataset 'meuse'
arc.write(file.path(fgdb_path, "meuse\\\\pts"), data=meuse, coords=c("x", "y", "elev"), shape_info=list(type='Point',hasZ=TRUE,WKID=28992))
data(meuse.riv, package="sp")
riv <- sp::SpatialPolygons(list(sp::Polygons(list(sp::Polygon(meuse.riv)),"meuse.riv")))
## write only geometry
arc.write(file.path(fgdb_path, "meuse\\\\riv"), coords=riv)

## write as table
arc.write(file.path(fgdb_path, "tlb"), data=list('f1'=c(23,45), 'f2'=c('hello', 'bob')))

## from scratch as feature class
arc.write(file.path(fgdb_path, "fc_pts"), data=list('data'=rnorm(100)),
          coords=list(x=runif(100,min=0,max=10),y=runif(100,min=0,max=10)),
          shape_info=list(type='Point'))

## write Raster
# make SpatialPixelsDataFrame
data(meuse.grid, package="sp")
sp::coordinates(meuse.grid) = c("x", "y")
sp::gridded(meuse.grid) <- TRUE
meuse.grid@proj4string=sp::CRS(arc.fromWktToP4(28992))

arc.write(file.path(fgdb_path, "meuse_grid"), meuse.grid)

## write using a RasterLayer object
r <- raster::raster(ncol=10, nrow=10)
raster::values(r) <- runif(raster::ncell(r))
arc.write(file.path(fgdb_path, "raster"), r)

}
\seealso{
\code{\link{arc.dataset-class}}, \code{\link{arc.open}}, \code{\link{arc.raster}}
}
\keyword{datasets}
\keyword{open}
\keyword{write}

