\name{arcgisbinding-package}
\docType{package}
\alias{arcgisbinding}
\alias{arcgisbinding-package}
\title{Bindings for ArcGIS}
\description{
Collection of classes and functons for loading, converting and exporting ArcGIS datasets and layers in R.
}
\section{Introduction}{
For a complete list of exported functions, use
\code{library(help = "arcgisbinding")}.
}

\section{References}{

\itemize{
 \item \href{https://r-forge.r-project.org/projects/rspatial/}{sp package} (Classes and Methods for Spatial Data)
 \item \href{https://github.com/r-spatial/sf/}{sf package (Simple Features for R)}
 \item \href{https://cran.r-project.org/package=raster}{raster package} (Geographic Data Analysis and Modeling)
 \item \href{https://cran.r-project.org/web/views/Spatial.html}{CRAN Task View: Analysis of Spatial Data}
}
}

\keyword{ package }
\keyword{ methods }
