% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.generics.R
\name{as.raster}
\alias{as.raster}
\title{Create RasterLayer or RasterBrick (raster package)}
\arguments{
\item{x}{\link{arc.raster-class} object}

\item{...}{TODO}
}
\value{
\code{RasterLayer} or \code{RasterBrick}
}
\description{
Create Rraster* object from arc.raster
TODO
}
\examples{
## convert arc.raster to Rasterlayer object
r.file <- system.file("pictures", "logo.jpg", package="rgdal")
r <- arc.raster(arc.open(r.file))
rx <- as.raster(r)
}

