% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA.classes.R
\name{dim}
\alias{dim}
\title{Dimensions of an object}
\arguments{
\item{x}{object \link{arc.raster-class} or \link{arc.datasetraster-class} object}
}
\value{
integer vector of \code{c(nrow, ncol, nband)}
dim,arc.raster-method
}
\description{
Get or set raster dimention
}

