% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc.data.R
\name{dplyr}
\alias{arrange.arc.data}
\alias{dplyr}
\alias{filter.arc.data}
\alias{group_by.arc.data}
\alias{mutate.arc.data}
\alias{ungroup.arc.data}
\title{dplyr support}
\usage{
filter.arc.data(.data, ..., .dots)

arrange.arc.data(.data, ..., .dots)

mutate.arc.data(.data, ..., .dots)

group_by.arc.data(.data, ..., add)

ungroup.arc.data(x, ...)
}
\description{
dplyr methods for arc.data objects: filter, arrange, mutate, select, group_by
}

