\name{internals}
\docType{methods}
\alias{pointData.arc.data}
\alias{polygonData.arc.data}
\alias{repr_html.arc.productinfo}
\alias{repr_html.arc.portalinfo}
\alias{repr_html.arc.data}
\alias{repr_html.arc.shapeinfo}
\alias{repr_html.leaflet}

\title{Jupyter notebook helpers}
\description{Internal jupyter notebook helpers}
\keyword{internal}
