% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA.classes.R
\name{names}
\alias{names}
\title{Get names}
\arguments{
\item{x}{object}
}
\value{
string vector
}
\description{
Return names of columns when \code{x} is \link{arc.table-class}.
Return band names when \code{x} is \link{arc.raster-class} or \link{arc.datasetraster-class}.
}

