% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arc.init.R
\name{arc.check_product}
\alias{arc.check_product}
\title{ArcGIS product and license information}
\usage{
arc.check_product()
}
\description{
Initialize connection to ArcGIS. Any script running directly from R (i.e.
without being called from a Geoprocessing script) should first call
\code{arc.check_product} to create a connection with ArcGIS. Provides
installation details on the version of ArcGIS installed that
\code{arcgisbinding} is communicating with.
}
\details{
Returned details include:
\itemize{
  \item Product: ArcGIS Desktop (i.e. ArcMap), or ArcGIS Pro. The name of the
        product connected.
  \item License level: Basic, Standard, or Advanced are the three licensing
        levels available. Each provides progressively more functionality
        within the software. See the "Desktop Functionality Matrix" link for
        details.
  \item Build number: The build number of the release being used.
        Useful in debugging and when creating error reports.
  \item DLL: The dynamic linked library (DLL) in use allowing
        ArcGIS to communicate with R.
}
}
\note{
Additional license levels are available on ArcGIS Desktop: Server,
       EngineGeoDB, and Engine. These license levels are currently
       unsupported by this package.
}
\section{References}{

 \href{http://www.esri.com/~/media/Files/Pdfs/library/brochures/pdfs/arcgis1021-desktop-functionality-matrix.pdf}{ArcGIS Desktop Functionality Matrix}
}
\examples{
info <- arc.check_product()
info$license # ArcGIS license level
info$version # ArcGIS build number
info$app # product name
info$dll # binding DLL in use
}

