% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc.data2sp.R
\name{arc.data2sp}
\alias{arc.data2sp}
\title{Convert an arc.dataframe or arc.raster object to an sp SpatialDataFrame object}
\usage{
arc.data2sp(x)
}
\arguments{
\item{x}{\code{data.frame} result of \code{\link{arc.select}} or \code{arc.raster}}
}
\description{
Convert an ArcGIS \code{data.frame} to the equivalent \code{sp} data frame
type. The output types that can be generated: \code{SpatialPointsDataFrame},
\code{SpatialLinesDataFrame}, or \code{SpatialPolygonsDataFrame}.
Convert an \code{arc.raster} object to a \code{SpatialGridDataFrame} object.
}
\examples{

require(sp)
d <- arc.select(arc.open(system.file("extdata", "ca_ozone_pts.shp", package="arcgisbinding")), 'ozone')
sp.df <- arc.data2sp(d)
\dontrun{spplot(sp.df)}
}

