% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.generics.R
\docType{methods}
\name{arc.metadata}
\alias{arc.metadata}
\title{Get metadata}
\usage{
arc.metadata(object)
}
\arguments{
\item{object}{\link{arc.dataset-class} object}
}
\description{
Get metadata
}

