% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.generics.R, R/arc.dataset.table.R
\docType{methods}
\name{arc.select}
\alias{arc.select}
\alias{arc.select,arc.table-method}
\title{Load dataset to data.frame}
\usage{
arc.select(object, fields, where_clause, selected, sr, ...)

\S4method{arc.select}{arc.table}(object, fields, where_clause, selected, sr,
  ...)
}
\arguments{
\item{object}{\link{arc.dataset-class} object}

\item{fields}{string, or list of strings, containing fields to include (default: all)}

\item{where_clause}{SQL where clause}

\item{selected}{use only selected records (if any) when dataset is a layer or
standalone table}

\item{sr}{transform geometry to Spatial Reference (default: object@sr)}

\item{...}{Additional arguements (currently ignored)}
}
\value{
\code{arc.select} returns a \code{data.frame} object (type of
\code{arc.data}).
}
\description{
Load dataset to a standard data frame.
}
\note{
If dataset includes the \code{arc.feature} attribute, the "shape" of class
\code{\link{arc.shape-class}} will be attached to the resulting
\code{data.frame} object.
}
\examples{

## read all fields
ozone.file <- system.file("extdata", "ca_ozone_pts.shp",
                          package="arcgisbinding")
d <- arc.open(ozone.file)
df <- arc.select(d, names(d@fields))
head(df, n=3)

## read 'name', 'fid' and geometry
df <- arc.select(d, c('fid', 'ozone'), where_clause="fid < 5")
nrow(df)

## transform points to "+proj=eqc"
df <- arc.select(d,"fid", where_clause="fid<5", sr="+proj=eqc")
arc.shape(df)

}
\seealso{
\code{\link{arc.open}}
}
\keyword{datasets}
\keyword{feature}
\keyword{open}
\keyword{select}
\keyword{table}

