% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc.shape.R
\name{arc.shape}
\alias{arc.shape}
\title{Get arc.shape object}
\usage{
arc.shape(df)
}
\arguments{
\item{df}{\code{arc.dataframe}}
}
\description{
Get \code{\link{arc.shape-class}} from \code{arc.dataframe}
}
\examples{

d <- arc.open(system.file("extdata", "ca_ozone_pts.shp",
                          package="arcgisbinding"))

df <- arc.select(d, 'ozone')
shp <- arc.shape(df)
length(shp$x)
}
\seealso{
\code{\link{arc.select}}
}

