% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc.data2sf.R
\name{arc.shape2sf}
\alias{arc.shape2sf}
\title{Convert Esri shape to sfc simple feature geometry}
\usage{
arc.shape2sf(shape)
}
\arguments{
\item{shape}{\code{\link{arc.shape-class}}}
}
\description{
Convert \code{\link{arc.shape-class}} to \code{sfc} simple feature geometry:
\code{POINT}, \code{MULTIPOINT}, \code{POLYGON}, \code{MULTIPOLYGON}, \code{LINESTRING}, \code{MULTILINESTRING}.
}
\examples{

require(sp)
d <- arc.select(arc.open(system.file("extdata", "ca_ozone_pts.shp", package="arcgisbinding")), 'ozone')
x <- arc.shape2sp(arc.shape(d))
\dontrun{plot(x)}
}
\seealso{
\code{\link{arc.shape}}
}

