% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.generics.R
\name{arc.sp2data}
\alias{arc.sp2data}
\title{deprecated Convert a sp SpatialDataFrame object to an arc.dataframe object}
\arguments{
\item{sp.df}{\code{SpatialPointsDataFrame}, \code{SpatialPolygonsDataFrame},
or \code{SpatialLinesDataFrame}}
}
\description{
deprecated Convert \code{sp} \code{SpatialPointsDataFrame}, \code{SpatialPolygonsDataFrame},
and \code{SpatialLinesDataFrame} objects to an ArcGIS-compatible \code{data.frame}.
}
\seealso{
\code{\link{arc.data2sp}}
}
\keyword{convert}
\keyword{sp}

