% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arc.sp2data.R
\name{arc.sp2data}
\alias{arc.sp2data}
\title{Convert a sp SpatialDataFrame object to an arc.dataframe object}
\usage{
arc.sp2data(sp.df)
}
\arguments{
\item{sp.df}{\code{SpatialPointsDataFrame}, \code{SpatialPolygonsDataFrame},
or \code{SpatialLinesDataFrame}}
}
\description{
Convert \code{sp} \code{SpatialPointsDataFrame}, \code{SpatialPolygonsDataFrame},
and \code{SpatialLinesDataFrame} objects to an ArcGIS-compatible \code{data.frame}.
}
\examples{
require(sp)
s1 <- Polygon(cbind(c(2,4,4,1,2),c(2,3,5,4,2)))
s2 <- Polygon(cbind(c(5,4,2,5),c(2,3,2,2)))
s3 <- Polygon(cbind(c(4,4,5,10,4),c(5,3,2,5,5)))
s4 <- Polygon(cbind(c(5,6,6,5,5),c(4,4,3,3,4)), hole = TRUE)
ss1 <- Polygons(list(s1), "s1")
ss2 <- Polygons(list(s2), "s2")
ss3 <- Polygons(list(s3, s4), "s3/4")
spp <- SpatialPolygons(list(ss1,ss2,ss3), 1:3)
sp.df <- SpatialPolygonsDataFrame(spp, data=data.frame(df=1:3),
                               match.ID = FALSE)
arc.df = arc.sp2data(sp.df)
arc.write(tempfile("sp_poly", fileext=".shp"), arc.df)
}
\seealso{
\code{\link{arc.data2sp}}
}
\keyword{convert}
\keyword{sp}

