% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-package.R
\docType{package}
\name{arcgisbinding}
\alias{arc}
\alias{arcgisbinding}
\alias{arcgisbinding-package}
\title{Bindings for ArcGIS}
\description{
This package provides classes for loading, converting and exporting ArcGIS
datasets and layers in R.
}
\section{Introduction}{
 For a complete list of exported functions, use
\code{library(help = "arcgisbinding")}.
}

\section{References}{

\itemize{
 \item \href{https://r-forge.r-project.org/projects/rspatial/}{sp package}
 \item \href{http://cran.r-project.org/web/views/Spatial.html}{CRAN Task View: Analysis of Spatial Data}
}
}

