% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA.classes.R
\docType{package}
\name{pixeltypes}
\alias{pixeltypes}
\alias{pixeltypes-package}
\title{Pixel types}
\description{
The pixel type of the referenced raster dataset.
}
\section{The types are}{

\itemize{
  \item "U1" 1 bit
  \item "U2" 2 bits
  \item "U4" 4 bits
  \item "U8" Unsigned 8 bit integers
  \item "S8" 8 bit integers
  \item "U16" Unsigned 16 bit integers
  \item "S16" 16 bit integers
  \item "U32" Unsigned 32 bit integers
  \item "S32" 32 bit integers
  \item "F32" Single precision floating point
  \item "F64" Double precision floating point
}
}

