% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA.classes.R
\docType{package}
\name{resampletypes}
\alias{resampletypes}
\alias{resampletypes-package}
\title{Resample types}
\description{
Resample types
}
\section{Supported}{

\itemize{
  \item "NearestNeighbor" Nearest neighbor assignment. This is the default
  \item "BilinearInterpolation" Bilinear interpolation
  \item "CubicConvolution" Cubic convolution
  \item "Majority" TODO
  \item "BilinearInterpolationPlus" TODO
  \item "BilinearGaussBlur" TODO
  \item "BilinearGaussBlurPlus" TODO
  \item "Average" TODO
  \item "Minimum" TODO
  \item "Average" TODO
  \item "VectorAverage" TODO
}
}

